refactor: move simulations functions in a file Simulation
