+++ src/Games/CENTIPEDE/Rules.hpp
@@ -1,0 +1,16 @@
+/*
+** EPITECH PROJECT, 2024
+** Arcade
+** File description:
+** centipede
+*/
+
+#pragma once
+
+#define SIZE_CENTIPEDE 9
+#define CENTIPEDE_SPEED 4
+#define GEN_CENTIPEDE 13
+#define WAVES_TO_WIN 20
+
+#define MAP_WIDTH 21
+#define MAP_HEIGHT 22
